Refactor rate limits and OFT admin controls
Moves rate-limit mutators in `RateLimiterModule` to internal functions and adds explicit `onlyOwner` admin wrappers across OFT/adapter contracts. The module now uses packed `uint112/uint32` config/state fields, updates storage slot layout, and tightens state decay/checkpoint logic and max-amount quoting behavior. Blocks negative supply for selected chains via the SupplyCheckingModule.

Also adds optional negative-supply support in `SupplyTrackingModule` with a guard that prevents `totalTransferFrom` from exceeding tracked supply unless explicitly allowed, plus updates related scripts, ABI config tooling, tests, and regenerated flat contracts.

+++ test/foundry/contracts/RateLimiterModuleTest.t.sol
@@ -230,9 +230,9 @@ contract RateLimiterModuleTest is Test {
     }
 
     function _config(
-        uint256 _outboundLimit,
+        uint112 _outboundLimit,
         uint32 _outboundWindow,
-        uint256 _inboundLimit,
+        uint112 _inboundLimit,
         uint32 _inboundWindow
     ) internal pure returns (RateLimiterModule.RateLimitConfig memory config) {
         config = RateLimiterModule.RateLimitConfig({
